Refactor: 채팅방 선택 버튼 컴포넌트화
1. 채팅방 이름과 참가자 수 넘겨주도록 했음
2. 채팅방 참가자 수가 100/50/0 명 이상일 때 각각 빨강/노랑/초록 표시되도록 했음
